requirements.txt
# Core dependencies (matches pyproject.toml)
PyJWT>=2.8
starlette>=0.27
